+++ tests/op/UtilsOP.sol
@@ -2,11 +2,11 @@
 
 pragma solidity 0.8.16;
 
-import "OpenZeppelin/openzeppelin-contracts@4.7.3/contracts/token/ERC20/IERC20.sol";
-import "OpenZeppelin/openzeppelin-contracts@4.7.3/contracts/token/ERC20/utils/SafeERC20.sol";
+import 'OpenZeppelin/openzeppelin-contracts@4.7.3/contracts/token/ERC20/IERC20.sol';
+import 'OpenZeppelin/openzeppelin-contracts@4.7.3/contracts/token/ERC20/utils/SafeERC20.sol';
 
-import "../Utils.sol";
-import "../../interfaces/op/IBankOP.sol";
+import '../Utils.sol';
+import '../../interfaces/op/IBankOP.sol';
 
 contract UtilsOP is Utils {
   using SafeERC20 for IERC20;
@@ -19,22 +19,18 @@ contract UtilsOP is Utils {
   address OP = 0x4200000000000000000000000000000000000042;
 
   function setUp() public virtual {
-    vm.label(bankAddress, "bankAddress");
+    vm.label(bankAddress, 'bankAddress');
 
-    vm.label(WETH, "WETH");
-    vm.label(DAI, "DAI");
-    vm.label(USDC, "USDC");
-    vm.label(OP, "OP");
-    vm.label(alice, "alice");
-    vm.label(liquidator, "liquidator");
+    vm.label(WETH, 'WETH');
+    vm.label(DAI, 'DAI');
+    vm.label(USDC, 'USDC');
+    vm.label(OP, 'OP');
+    vm.label(alice, 'alice');
+    vm.label(liquidator, 'liquidator');
   }
 
-  function balanceOf(address token, address user)
-    internal
-    view
-    returns (uint256)
-  {
-    uint256 balance = IERC20(token).balanceOf(user);
+  function balanceOf(address token, address user) internal view returns (uint) {
+    uint balance = IERC20(token).balanceOf(user);
     if (token == WETH) {
       return balance + address(user).balance;
     }
@@ -77,7 +73,7 @@ contract UtilsOP is Utils {
     IBankOP _bank,
     address _user,
     address _token,
-    uint256 _limit,
+    uint _limit,
     address _origin
   ) internal {
     IBankOP.CreditLimit[] memory creditLimits = new IBankOP.CreditLimit[](1);
